Fixed delete course API && Corrected few pros-cons && added some conditions at course module

+++ src/main/java/com/teaminvincible/ESchool/CourseModule/ServiceImplementation/CourseServiceImplementation.java
@@ -13,11 +13,11 @@ import com.teaminvincible.ESchool.Utility.CodeGenerator;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
@@ -32,19 +32,6 @@ public class CourseServiceImplementation implements CourseService {
     private UserDescriptionService userDescriptionService;
 
 
-    public Course findCourseById(String userID, String courseId) throws CustomException{
-
-        Specification<Course> spec = Specification.where(
-                CourseSpecification.findCourseByCourseOwner(userID)
-                        .and(CourseSpecification.findCourseByCourseId(courseId))
-        );
-
-        Course course = courseRepository.findAll(spec).get(0);
-
-        if(course != null)return course;
-        throw new CustomException(HttpStatus.NOT_FOUND,"No course found with id: "+courseId);
-    }
-
     public Course findCourseById(String courseId) throws CustomException{
 
         Optional<Course> course = courseRepository.findById(courseId);
@@ -55,19 +42,16 @@ public class CourseServiceImplementation implements CourseService {
 
     @Override
     public Course createCourse(String userId, Course course) throws CustomException {
-        /**
-         *  Before saving we have to use "UserDescriptionService" to validate the user, exist, active or not
-         */
-
-        Course course1 = null;
-
-        if(course.getCourseJoiningCode() == null)
-            course.setCourseJoiningCode(CodeGenerator.generateCourseJoinCode());
 
         UserDescription userDescription = userDescriptionService.getUserDescription(userId);
 
         if(userDescription.getRole() == Role.STUDENT)
             throw new CustomException(HttpStatus.BAD_REQUEST,"Students are not allowed to create a course! They can only join.");
+
+        Course course1;
+
+        if(course.getCourseJoiningCode() == null)
+            course.setCourseJoiningCode(CodeGenerator.generateCourseJoinCode());
 
         course.setCourseOwner(userDescription);
 
@@ -84,28 +68,21 @@ public class CourseServiceImplementation implements CourseService {
     public Course joinCourse(String userId, String joiningCode) throws CustomException {
         UserDescription userDescription = userDescriptionService.getUserDescription(userId);
 
-        System.out.println("User Role: "+userDescription.getRole());
-
         if(userDescription.getRole() == Role.TEACHER)
             throw new CustomException(HttpStatus.BAD_REQUEST,"Teachers are not allowed to join a course! They can only create course.");
 
         Course course = courseRepository.findBycourseJoiningCode(joiningCode);
 
-        System.out.println("User Role: "+userDescription.getRole());
-
-
         if(course == null)
-            throw new CustomException(HttpStatus.BAD_REQUEST,"No course found! Please provide a valid course joining code.");
+            throw new CustomException(HttpStatus.BAD_REQUEST,"No course found! Please provide a valid course Joining Code.");
 
-        // Check user exist in this course or not
+        if(userDescription.checkIfUserAlreadyEnrolledThisCourse(course))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"Opps! Already enrolled in this course.");
 
         Set<Course> courses = userDescription.getCourses();
         courses.add(course);
 
         userDescription.setCourses(courses);
-
-        System.out.println("User Role: "+userDescription.getRole());
-
 
         userDescriptionService.updateUserDescription(userDescription);
 
@@ -117,19 +94,30 @@ public class CourseServiceImplementation implements CourseService {
 
         Course courseToRemove= findCourseById(courseId);
 
+        if(Objects.isNull(courseToRemove))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"No course found with course-id: "+courseId);
+
         UserDescription userDescription = userDescriptionService.getUserDescription(userId);
+
+        if(!userDescription.checkIfUserAlreadyEnrolledThisCourse(courseToRemove))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"Opps! You're not enrolled in this course.");
 
         userDescription.removeCourse(courseToRemove);
 
         userDescriptionService.updateUserDescription(userDescription);
-
 
         return "Successfully un-enrolled from course: "+courseToRemove.getCourseTitle();
     }
 
     @Override
     public Course updateCourse(String userId, Course course) throws CustomException {
-        Course oldCourse = findCourseById(userId, course.getCourseId());
+        Course oldCourse = findCourseById(course.getCourseId());
+
+        if(Objects.isNull(oldCourse))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"No course found with course-id: "+course.getCourseId());
+
+        if(!oldCourse.getCourseOwner().getUserId().equals(userId))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"Sorry! You're not the owner of this course.");
 
         Course updatedCourse = new Course();
 
@@ -143,6 +131,14 @@ public class CourseServiceImplementation implements CourseService {
     @Override
     public String deleteCourse(String userId, String courseId) throws CustomException {
 
+        Course course = findCourseById(courseId);
+
+        if(Objects.isNull(course))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"No course found with course-id: "+course.getCourseId());
+
+        if(!course.getCourseOwner().getUserId().equals(userId))
+            throw new CustomException(HttpStatus.BAD_REQUEST,"Sorry! You're not the owner of this course.");
+
         try{
             courseRepository.deleteById(courseId);
         }catch (Exception ex){
@@ -153,9 +149,10 @@ public class CourseServiceImplementation implements CourseService {
     }
 
     @Override
-    public Set<Course> getAllCourseOfAUser(String userId) throws CustomException {
+    public Set<Course> getAllCourseOfATeacher(String userId) throws CustomException {
 
         Set<Course> courseSet = new HashSet<>();
+
         courseSet.addAll( courseRepository.findAll(CourseSpecification.findCourseByCourseOwner(userId)));
 
         return courseSet;
